+++ ElvUI_MerathilisUI/Settings/Profile.lua
@@ -1409,6 +1409,27 @@ P.buffReminder = {
 	},
 }
 
+P.battleRes = {
+	enable = false,
+	visibility = "MPLUS_AND_RAID", -- MPLUS_AND_RAID, MPLUS, RAID
+	displayMode = "ICON", -- ICON, TEXT
+	iconSize = 40,
+	desaturate = true,
+	countFont = {
+		name = I.Fonts.Primary,
+		size = 14,
+		style = "SHADOWOUTLINE",
+	},
+	countColor = { r = 1, g = 1, b = 1 },
+	timeFont = {
+		name = I.Fonts.Primary,
+		size = 14,
+		style = "SHADOWOUTLINE",
+	},
+	timeColor = { r = 1, g = 1, b = 1 },
+	debug = false,
+}
+
 P.movementAlert = {
 	enable = false,
 	combatOnly = false,